# extras/mystore-backend/openapi.yaml
openapi: 3.0.3
info:
  title: MyStore Backend API
  description:
    "MyStore Backend API. The API version is selected dynamically with
    the `{version}` path parameter (`v1` or `v2`). Bearer authentication is optional
    in the OpenAPI definition: clients may send an application JWT when available."
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/{version}/favorites:
    get:
      tags:
        - Favorites
      summary: List favorites
      responses:
        "200":
          $ref: "#/components/responses/FavoritesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
      operationId: getFavorites
      security:
        - bearerAuth: []
        - {}
      description: "If a token is provided, send it as **`Authorization: Bearer <application-jwt>`**."
    post:
      tags:
        - Favorites
      summary: Add favorite
      responses:
        "200":
          $ref: "#/components/responses/NullDataResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
      operationId: addFavorite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddFavoriteRequest"
      security:
        - bearerAuth: []
        - {}
      description: "If a token is provided, send it as **`Authorization: Bearer <application-jwt>`**."
    parameters:
      - $ref: "#/components/parameters/ApiVersion"
  /api/{version}/favorites/{productId}:
    delete:
      tags:
        - Favorites
      summary: Remove favorite
      responses:
        "200":
          $ref: "#/components/responses/NullDataResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
      operationId: removeFavorite
      parameters:
        - $ref: "#/components/parameters/FavoriteProductId"
      security:
        - bearerAuth: []
        - {}
      description: "If a token is provided, send it as **`Authorization: Bearer <application-jwt>`**."
    parameters:
      - $ref: "#/components/parameters/ApiVersion"
...
  /api/{version}/orders:
    post:
      tags:
        - Orders
      summary: Create order
      responses:
        "201":
          $ref: "#/components/responses/OrderResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      security:
        - bearerAuth: []
        - {}
      description: "If a token is provided, send it as **`Authorization: Bearer <application-jwt>`**."
    get:
      tags:
        - Orders
      summary: List my orders
      responses:
        "200":
          $ref: "#/components/responses/OrdersResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
      operationId: getOrders
      security:
        - bearerAuth: []
        - {}
      description: "If a token is provided, send it as **`Authorization: Bearer <application-jwt>`**."
    parameters:
      - $ref: "#/components/parameters/ApiVersion"
...
components:
  parameters:
...
    FavoriteProductId:
      name: productId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ApiVersion:
      name: version
      in: path
      required: true
      description: API version. Use `v1` or `v2`.
      schema:
        type: string
        enum:
          - v1
          - v2
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: MyStore application JWT. Used by v2 protected endpoints.
  schemas:
    Price:
      type: object
      required:
        - amount
        - currency
      properties:
        amount:
          type: number
          format: double
          example: 19.99
        currency:
          type: string
          example: USD
        discount_percent:
          type: number
          format: double
          example: 10
...
    FavoritesResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          items:
            type: string
            format: uuid
    AddFavoriteRequest:
      type: object
      required:
        - product_id
      properties:
        product_id:
          type: string
          format: uuid
...
    CreateOrderRequest:
      type: object
      required:
        - cart_id
        - payment_id
        - payment_method_id
      properties:
        cart_id:
          type: string
          format: uuid
        payment_id:
          type: string
        payment_method_id:
          type: string
    OrderLineItem:
      type: object
      properties:
        product_id:
          type: string
          format: uuid
        quantity:
          type: integer
        purchase_price:
          $ref: "#/components/schemas/Price"
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        placed_at:
          type: string
          format: date-time
        payment_id:
          type: string
        payment_method_id:
          type: string
        line_items:
          type: array
          items:
            $ref: "#/components/schemas/OrderLineItem"
    OrderResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/Order"
    OrdersResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - orders
          properties:
            orders:
              type: array
              items:
                $ref: "#/components/schemas/Order"
    NullDataResponse:
      type: object
      required:
        - data
      properties:
        data:
          nullable: true
          example: null
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
...
  responses:
    BadRequest:
      description: Invalid input or workflow state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated user lacks ownership or permission
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource or route does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    FavoritesResponse:
      description: Favorites
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/FavoritesResponse"
    NullDataResponse:
      description: NullData
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/NullDataResponse"
...
    OrderResponse:
      description: Order
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OrderResponse"
    OrdersResponse:
      description: Orders
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OrdersResponse"
...